test(api): move CAB-2195 regression test to test_regression_*.py file
regression-guard.yml matches on file name pattern test_regression_*.py;
the test in test_keycloak_service.py wasn't recognised. Move it to a
dedicated tests/test_regression_cab_2195.py without changing the
assertions.
