+++ control-plane-api/tests/test_keycloak_service.py
@@ -38,22 +38,6 @@ class TestConnect:
         ):
                 await svc.connect()
 
-    async def test_regression_cab_2195_connect_uses_user_realm_name(self):
-        # Authenticate against master (where the admin user lives) and operate on
-        # the configured realm. Mutating connection.realm_name post-init breaks
-        # token refresh because python-keycloak hits the token endpoint of the
-        # mutated realm — where `admin` does not exist — yielding 401
-        # invalid_grant on every refresh.
-        from src.config import settings
-
-        svc = KeycloakService()
-        with patch("src.services.keycloak_service.KeycloakOpenIDConnection") as mock_conn, \
-             patch("src.services.keycloak_service.KeycloakAdmin"):
-            await svc.connect()
-            kwargs = mock_conn.call_args.kwargs
-            assert kwargs["realm_name"] == settings.KEYCLOAK_REALM
-            assert kwargs["user_realm_name"] == "master"
-
 
 class TestDisconnect:
     async def test_disconnect_clears_admin(self, svc):
